+++ src/target_reduction.cpp
@@ -1,18 +1,17 @@
 #include "target_reduction.hpp"
 
-
 namespace sep
 {
-TargetReduction::TargetReduction(
-    sep::UnitType* unit_type,
-    unsigned int value) 
-{
-    unit_type_ = unit_type;
-    value_ = value;
-}
+    TargetReduction::TargetReduction(
+        sep::UnitType* unit_type,
+        uint16_t value) 
+    {
+        unit_type_ = unit_type;
+        value_ = value;
+    }
 
-TargetReduction::~TargetReduction() 
-{
-    delete unit_type_;
-}
-}
+    TargetReduction::~TargetReduction() 
+    {
+        delete unit_type_;
+    }
+}  // namespace sep
